[DP-12820] cleanup pipeline by removing unnecessary secrets and commands (#358)
This is a followup to DP-12820 to clean up semaphore pipelines. This change removes some secrets that have been
deactivated as well as removes some commands that are set automatically by the semaphore agent, and replaces
two lines make install-vault and mk-include/bin/vault-setup with . vault-setup.

The same [change](https://github.com/confluentinc/cc-service-bot/pull/471/files) was made for managed semaphore pipelines, this automated PR
should match the changes made in the managed semaphore pipelines.

This should not affect your pipelines. If the commands match what was made in the managed semaphore pipelines,
and the pr build passes, please merge this change.