FLINK-34007 (#531)
* [hotfix][test] Moves ConfigMap lifecycle management into @BeforeEach/@AfterEach methods

* [hotfix][test] Makes ManuallyTriggeredScheduledExecutorService#execute rely on a BlockingQueue

This allows us to wait for tasks to "arrive".

* [hotfix][test] Refactors TestingLeaderCallbackHandler to allow async calls

This way, we can use FlinkAssertions#assertThatFuture and use assertion messages instead of comments.

* [hotfix][core] Makes ExecutorUtils#gracefulShutdown return any outstanding tasks

* [FLINK-34333][k8s] fabric8io LeaderElector is created with every new #run() call (cherry-picked from FLINK-34007)

v5.12.4 allowed us to reuse the LeaderElector. With v6.6.2 (fabric8io/kubernetes-client#4125) this behavior changed. One LeaderElector can only be used until the leadership is lost.
An ITCase is added to cover the scenario where the leadership is lost.

* [FLINK-34333][k8s] Upgrade k8s client to v6.9.2 to cover client issue #5464 (cherry-picked from FLINK-34007)

https://github.com/fabric8io/kubernetes-client/issues/5463

The change in KubernetesClientTestBase was necessary because of the change [#5369](https://github.com/fabric8io/kubernetes-client/pull/5369) that ended up in v6.9.0 as well.

* fixup! [FLINK-34333][k8s] fabric8io LeaderElector is created with every new #run() call (cherry-picked from FLINK-34007)

* fixup! [hotfix][test] Makes ManuallyTriggeredScheduledExecutorService#execute rely on a BlockingQueue

---------

Co-authored-by: Matthias Pohl <matthias.pohl@aiven.io>