Add integration-impl-deps (#543)
Adding integration-impl-deps module to be sure that we're capturing all the transitive dependencies correctly when we build the integration-impl jar in the Docker build.

Unfortunately, we have to merge this PR with both the new and old approaches so that we can publish the new module without breaking the runtime. We have to publish the new module so that we can build the Docker project with the dependency.

There will be a follow-up PR to remove the integration-impl module after it's been added to the Docker build.