[FLINKCC-681] Limit Local Aggregation buffers
The buffer holds records for aggregation in memory and is flushed downstream on:
1) checkpoints, 2) watermarks, or 3) when it is full.
If it's too big (even if watermarks are progressing),
there might be much more data to flush than the memory available (classic flatMap problem).
That might lead to back-pressure and hard-blocking the task thread.

This change limits the buffer and the number of records.