[FSTORAGE-72] Remove cc-flink-managed-table module from flink repository (#595)
This change removes cc-flink-managed-table module from flink repository. Also, remove its references in this repository. This module is being moved to kdb repository.

Also, removed CompiledPlanJobGraphGeneratorV3ImplTest. It fails currently because we are removing
cc-flink-managed-table module as part of this PR and this test depends on that module.
The test is also present at cc-flink-docker/cc-flink-runtime-integration-impl module.